requirements.txt
google-generativeai>=0.7.0
python-dotenv>=1.0.1
moviepy==1.0.3
openai-whisper>=20231117
torch>=2.1.0
gTTS>=2.5.1
pydub>=0.25.1
numpy>=1.26.0

